# backend/baangs_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Baangs FSM API",
    "description": "CCTV Field Service Management System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/export/jobs.csv": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export Jobs Csv",
        "operationId": "export_jobs_csv_export_jobs_csv_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/export/customers.csv": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export Customers Csv",
        "operationId": "export_customers_csv_export_customers_csv_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/export/billing.csv": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export Billing Csv",
        "operationId": "export_billing_csv_export_billing_csv_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/export/attendance.csv": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export Attendance Csv",
        "operationId": "export_attendance_csv_export_attendance_csv_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/export/all.json": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export All Json",
        "operationId": "export_all_json_export_all_json_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}